ATR-419: refactored out BoundType enum into separate file
