+++ src/Acuminator/Acuminator.Utilities/Roslyn/PXFieldAttributes/AttributeInformation.cs
@@ -11,20 +11,6 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 namespace Acuminator.Utilities.Roslyn.PXFieldAttributes
 {
 	/// <summary>
-	/// Helper used to classify Acumatica attributes.
-	/// </summary>
-	/// <remarks>
-	/// By Acumatica atribute we mean an attribute derived from PXEventSubscriberAttribute.
-	/// </remarks>
-	
-	public enum BoundType
-	{
-		Unknown = 0,
-		Unbound = 1,
-		DbBound = 2,
-		NotDefined = 3
-	}
-	/// <summary>
 	/// Helper used to retrieve information about the Acumatica attributes.
 	/// </summary>
 	/// <remarks>
